sched: fair: Avoid little cpus due to sync, prev bias
Important threads can get forced to little cpu's
when the sync or prev_bias hints are followed
blindly. This patch adds a check to see whether
those paths are forcing the task to a cpu that
has less capacity than other cpu's available for
the task. If so, we ignore the sync and prev_bias
and allow the scheduler to make a free decision.

Bug: 117438867
Signed-off-by: Miguel de Dios <migueldedios@google.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>