sched: Mitigate non-boosted tasks preempting boosted tasks
Currently when a boosted task is scheduled we use prefer_idle to try and
get it to an idle core. Once it's scheduled, there is a possibility we
can schedule a non-boosted task on the same core where the boosted task
is running on. This change aims to mitigate that possibility by checking
if the core we're targeting has a boosted task and if so, use the next
best idle core instead.

Bug: 131626264
Signed-off-by: Miguel de Dios <migueldedios@google.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>