sched/fair: Let scheduler skip util checking if CPU is idle
Current cpu util includes util of runnable tasks plus the recent utilization
of currently non-runnable tasks, so it may return a non-zero value even
there is no task running on a cpu. When scheduler is selecting a cpu for
a task, it will check if cpu util is over its capacity, what could
happen is that it will skip a cpu even it is idle, so let scheduler
skip util checking if the task perfers idle cpu and the cpu is idle.

Bug: 133284637
Test: cpu selected as expected
Signed-off-by: Rick Yiu <rickyiu@google.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>